+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Our versioning strategy is as follows:
 
 ## 22.5.0
 
+### 🎉 New Features & Improvements
+
+* `[sitecore-jss-angular]` Support `CanActivate` `RedirectCommand` API ([#2029](https://github.com/Sitecore/jss/pull/2029))
+* `[sitecore-jss-angular]` `CanActivate` and `CanResolve` now additionally accept `router` as a parameter ([#2029](https://github.com/Sitecore/jss/pull/2029))
+
 ### 🛠 Breaking Change
 
 * Upgrade Angular to v18 ([#2029](https://github.com/Sitecore/jss/pull/2029))